BeeMonitor
The honeybees are dying - and we don't really know why. On normal years beekeepers might expect to lose around 10% of their colony. Over the last years mortality rates has changed up to 75%.

The health of colony can be determined based on a few factors. 
- Temperature is an early indicator of whether or not the bees are about to swarm. 
- Humidity is another indicator of bee health. If the area is too damp, it can damage the hive.

So here is the idea !
I want to build a data logger for the hives. I like to record the temperature inside as well as outside temperature and humidity

http://forum.arduino.cc/index.php?topic=303854.0